requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf_yasg==1.21.7
celery==5.3.4
django-environ==0.11.2
mysqlclient==2.2.0
rabbitmq==0.2.0